+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## 1.2.2
+
+* Few Adjustments
+
+
+## 1.2.1
+
+* Minor changes in Color Lightness.
+
+
+## 1.2.0
+
+* Release Version with functioning Colors.
+
+
+## 1.1.6
+
+* Fix Bug
+
+
+## 1.1.5
+
+* Change Focus Color
+
+
+## 1.1.4
+
+* Fix Bugs concering constant expressions
+
+
+## 1.1.3
+
+* Change Border Color to main Color
+
+
 ## 1.1.2
 
 * Update Dark Text Theme
@@ -6,6 +41,7 @@
 ## 1.1.1
 
 * Update stuff
+
 
 ## 1.1.0